@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0"
 
 dependencies:
   flutter:
@@ -27,7 +27,8 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.0
-  english_words: ^3.1.5            
+  english_words: ^3.1.5
+  flutter_local_notifications: ^9.4.0            
 
 dev_dependencies:
   flutter_test:
