@@ -1,18 +1,22 @@
+import 'NotificationService.dart';
 import 'package:flutter/material.dart';
 
+
 class MainDashboard extends StatelessWidget {
-  const MainDashboard({Key key}) : super(key: key);
+  MainDashboard({Key? key}) : super(key: key);
+
+ NotificationService ns = NotificationService();
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: const Text('Second Route'),
+        title: const Text('MainDashboard'),
       ),
       body: Center(
         child: ElevatedButton(
           onPressed: () {
-            Navigator.pop(context);
+            ns.showNotification();
           },
           child: const Text('Go back!'),
         ),
