@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'MainDashboard.dart';
 
 class LoginPage extends StatefulWidget {
-  LoginPage({Key key, this.title}) : super(key: key);
-  final String title;
+  LoginPage({Key? key, this.title}) : super(key: key);
+  final String? title;
   @override
   LoginPageState createState() => LoginPageState();
 }
@@ -42,7 +42,7 @@ class LoginPageState extends State<LoginPage> {
         onPressed: () {
           Navigator.push(
             context,
-            MaterialPageRoute(builder: (context) => const MainDashboard()),
+            MaterialPageRoute(builder: (context) => MainDashboard()),
           );
         },
         child: Text("Login",
